+++ lib/canvas-workflow-jekyll/file.rb
@@ -1,5 +1,36 @@
 require 'pandarus'  # canvas lms api
 
+# monkey-patch incorrect implementation
+module Pandarus
+  class Client
+    def resolve_path_courses(course_id, path, opts={})
+      query_param_keys = [
+      ]
+
+      form_param_keys = [
+      ]
+
+      # verify existence of params
+      raise "course_id is required" if course_id.nil?
+      raise "path is required" if path.nil?
+      # set default values and merge with input
+      options = underscored_merge_opts(opts,
+        :course_id => course_id,
+        :path => path
+      )
+
+      # resource path
+      path = path_replace("/v1/courses/{course_id}/folders/by_path/{path}/",
+        :course_id => course_id, :path => path)
+      headers = nil
+      form_params = select_params(options, form_param_keys)
+      query_params = select_query_params(options, query_param_keys)
+
+      RemoteCollection.new(connection, Folder, path, query_params)
+    end
+  end
+end
+
 module Canvas
   module Workflow
     module Jekyll
@@ -15,33 +46,6 @@ module Canvas
           prefix = config['canvas']['prefix'] + '/api'
           course = config['canvas']['course']
           client = Pandarus::Client.new(prefix: prefix, token: token)
-
-          # patch incorrect implementations
-          client.define_singleton_method(:resolve_path_courses) do |course_id, path, opts={}|
-            query_param_keys = [
-            ]
-
-            form_param_keys = [
-            ]
-
-            # verify existence of params
-            raise "course_id is required" if course_id.nil?
-            raise "path is required" if path.nil?
-            # set default values and merge with input
-            options = underscored_merge_opts(opts,
-              :course_id => course_id,
-              :path => path
-            )
-
-            # resource path
-            path = path_replace("/v1/courses/{course_id}/folders/by_path/{path}/",
-              :course_id => course_id, :path => path)
-            headers = nil
-            form_params = select_params(options, form_param_keys)
-            query_params = select_query_params(options, query_param_keys)
-
-            Pandarus::RemoteCollection.new(connection, Pandarus::Folder, path, query_params)
-          end
 
           dir     = File.dirname(@path)
           file    = File.basename(@path)