+++ lib/canvas-workflow-jekyll/assignment.rb
@@ -1,0 +1,28 @@
+module Canvas
+  module Workflow
+    module Jekyll
+      class AssignmentTag < Liquid::Tag
+        def initialize(tag_name, text, tokens)
+          super
+        end
+
+        def render(context)
+          config = context.registers[:site].config['canvas']
+          title  = context.environments.first['page']['title']
+          course = config['course']
+          client = Canvas::Client.new(config)
+
+          assignments = client.list_assignments(course, :search_term => title).to_a
+
+          raise ArgumentError.new("Assignment does not exist") if assignments.empty?
+
+          # return the first, which /should/ be the shortest length string, so
+          # first lexicographically
+          assignments.first[:id]
+        end
+      end
+    end
+  end
+end
+
+Liquid::Template.register_tag('assignment', Canvas::Workflow::Jekyll::AssignmentTag)