+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Enhancements
+* Moved 'development' dependencies to Gemfile and listed bundler as only
+  development dependency in gemspec.
+
 ## [0.3.0](https://github.com/jiverson002/canvas-workflow-jekyll/releases/tag/0.3.0)
 * Renamed project to canvas-workflow-jekyll.
 * Incorporated the former standalone gem jekyll-canvas-tags into this project.