+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Enhancements
 * Moved 'development' dependencies to Gemfile and listed bundler as only
   development dependency in gemspec.
+* Made gist iframe width 100%.
 
 ## [0.3.0](https://github.com/jiverson002/canvas-workflow-jekyll/releases/tag/0.3.0)
 * Renamed project to canvas-workflow-jekyll.