+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Moved 'development' dependencies to Gemfile and listed bundler as only
   development dependency in gemspec.
 * Made gist iframe width 100%.
+* Refactored monkey-patch of resolve\_path\_courses in
+  Canvas::Workflow::Jekyll::FileTag out of FileTag class.
 
 ## [0.3.0](https://github.com/jiverson002/canvas-workflow-jekyll/releases/tag/0.3.0)
 * Renamed project to canvas-workflow-jekyll.