+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Made gist iframe width 100%.
 * Refactored monkey-patch of resolve\_path\_courses in
   Canvas::Workflow::Jekyll::FileTag out of FileTag class.
-* Upgraded to canvas-workflow-travis v0.10.0.
+* Upgraded to canvas-workflow-travis v0.11.0.
 
 ## [0.3.0](https://github.com/jiverson002/canvas-workflow-jekyll/releases/tag/0.3.0)
 * Renamed project to canvas-workflow-jekyll.