+++ CHANGELOG.md
@@ -5,7 +5,7 @@
   development dependency in gemspec.
 * Made gist iframe width 100%.
 * Refactored monkey-patch of resolve\_path\_courses in
-  Canvas::Workflow::Jekyll::FileTag out of FileTag class.
+  Canvas::Workflow::Jekyll::FileTag out of FileTag class into pandarus.rb.
 * Upgraded to canvas-workflow-travis v0.12.0.
 
 ## [0.3.0](https://github.com/jiverson002/canvas-workflow-jekyll/releases/tag/0.3.0)